Tests/LibWeb/Text/expected/HTML/StructuredClone-object-primitives.txt

true
false
123
123.456
This is a String object
1692748800000
